Move /deploy-config to /config
In order to consolidate endpoints that manipulate CDN configuration
the previously used endpoint /deploy-config is moved to /config enpoint.

The old endpoint is defined as deprecated and calls the `config_post()`
directly providing the same fuctionality as before.

The new endpoint provides the same functionality.
